+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 - Use react-oidc-context for authentication.
 - Use OpenID Connect "sub" claim as user identifier.
 - Expanded API health checks.
+- Authenticated Users are now registed in the database.
 
 ## v1.0.93 - 2024-05-14